Refactor suggtool into a base command and package
